chore:room과 users Repository의 자료형 오류로 수정

+++ src/main/java/com/bfriend/bfriend/room/RoomRepository.java
@@ -2,7 +2,7 @@ package com.bfriend.bfriend.room;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RoomRepository extends JpaRepository<Room, Integer> {
-    public Room findByRid(Integer rid);
+public interface RoomRepository extends JpaRepository<Room, Long> {
+    public Room findByRid(Long rid);
 
 }
